Break line
A digital ATM card is a virtual representation of a physical ATM or debit card, designed to be used primarily through smartphones, tablets, or smartwatches. It functions similarly to a traditional ATM card, allowing users to access their bank accounts to withdraw cash, make deposits, and perform other transactions at ATMs. However, it can also be used for online transactions and at point-of-sale (POS) terminals that support contactless payments.